Try to separate tests out to monitor_test.py
